Implement ortho, quat_from and rigid_inv
This commit toys with the idea of using 3x4 matrices instead of 4x4
matrices for a 25% savings in memory (and perhaps speed when pushing
many transforms to the GPU).

Rigid (rotation and translation only) transform inverse are
simpler/faster to calculate; this is lacking in DXM.  Implement it.
This is sometimes called affine inverse e.g. `glm::affineInverse`.

Finding an arbitrary vector’s orthogonal and calculating a quaternion
that rotates a vector into another are other useful utilities in games
and simulations; both branch and branchless versions are implemented.